requirements.txt
python-binance==1.0.16
python-telegram-bot==13.15
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
requests==2.29.0
mplfinance==0.12.9b0
python-dotenv==1.0.0
schedule==1.2.0